refactor: ViewController 파일 분리 - WeatherCollectionView.swift - WeatherCollectionViewCell.swift
